backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
flower==2.0.1
websockets==12.0
